+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+- Added optional `centered` prop to `Tabs` for icon alignment
 ## [0.22.2]
 - Improved visual bug involving skrinking / resizing for `Table` and `Accordion`